requirements.txt
flask
flask-socketio
pymongo
python-dotenv
groq
gunicorn
eventlet
